refactor: 파일 컨벤션 및 DB Column 컨벤션 수정
