+++ src/config/db.config.ts
@@ -3,6 +3,7 @@ import { ConfigService } from '@nestjs/config';
 import { TypeOrmModuleOptions, TypeOrmOptionsFactory } from '@nestjs/typeorm';
 import { Content } from '@src/entities/content.entity';
 import { Tag, TagName } from '@src/entities/tag.entity';
+import { SnakeNamingStrategy } from 'typeorm-naming-strategies';
 
 @Injectable()
 class TypeOrmConfigService implements TypeOrmOptionsFactory {
@@ -19,6 +20,7 @@ class TypeOrmConfigService implements TypeOrmOptionsFactory {
       synchronize: true,
       logging: process.env.NODE_ENV === 'dev',
       entities: [Tag, TagName, Content],
+      namingStrategy: new SnakeNamingStrategy(),
     };
   }
 }
