+++ src/entities/base.entity.ts
@@ -17,9 +17,9 @@ export class Base extends BaseEntity {
   @PrimaryGeneratedColumn()
   seq: number;
 
-  @CreateDateColumn({ name: 'created_at' })
+  @CreateDateColumn()
   createdAt!: Date;
 
-  @UpdateDateColumn({ name: 'updated_at' })
+  @UpdateDateColumn()
   updatedAt!: Date;
 }
